+++ CHANGELOG.md
@@ -3,9 +3,21 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- the topmost header version must be set manually in the VERSION file -->
+
+### Version 1.1.32 2019-11-20
+- Improved Sql-logging
+
+### Version 1.1.31 2019-10-30
+- Added Disabled as Default Enum value in TriggersAndViewsGeneration enum
+
+### Version 1.1.30 2019-10-30
+- Implemented a mechanism for handling configuration-reload of changed config-files
+- Added more Code documentation 
+
 ### Version 1.1.29 2019-10-24
 - Improved LogFileAppendFluentMigratorLoggerProvider
   Will try 10 times with wait of 250ms to create StreamWriter if IOException is thrown 
+
 ### Version 1.1.28 2019-09-27
 - Fixed  FluentDbTools.Common.Abstractions.Extensions.ExtractSqlStatements
 - Added test to verify that ExtractSqlStatements do the work